Fixes #82.
When Bitfinex disconnects the websocket we attempt to reconnect and
resubscribe to our channels, but the channel name gets rewritten from
(for example) "ticker" to "ticker-BTCUSD" after the first time which
results in Bitfinex saying it doesn't know what channel we're trying
to subscribe to. This change will rewrite the channel name to remove
the currency pair and allow the resubscribe to succeed.